+++ CHANGES.rst
@@ -8,7 +8,8 @@ Changes
 
 Changes:
 --------
-- No change.
+- Add ``weaver.wps.utils.get_wps_client`` function to handle the creation of ``owslib.wps.WebProcessingService`` client
+  with appropriate request options configuration from application settings.
 
 Fixes:
 ------